Refactor: `sendTransaction` structure
